all: use CINT for dict generation (ROOT-5)
- consolidate dict generation across ROOT versions (ROOT-5, ROOT-6)
- generate dict for []string (ie: golang::goslice<golang::gostring>)
