+++ bindings/src/gen-goedm-dict-root5.go
@@ -2,5 +2,5 @@
 
 package main
 
-//go:generate /bin/cp -f goedm_dict_root5.icc goedm_dict.cxx
+//go:generate rootcint -f goedm_dict.cxx -c -p goedm.h goedm_linkdef.h
 //go:generate touch goedm_dict_rdict.pcm
