+++ bindings/src/gen-goedm-dict-root6.go
