+++ bindings/src/goedm.h
@@ -1,29 +1,28 @@
-
 #ifndef GOEDM_GOEDMDICT_H
 #define GOEDM_GOEDMDICT_H 1
 
+#include <stdint.h> // for (u)intXXX_t
+
 namespace golang {
+
   struct gostring {
     int   Len;
     char *Data; //[Len]
   };
 
-  template <typename T>
-  struct goslice {
-    int  Len;
-    int  Cap;
-    T   *Data; //[Len]
-  };
-
+  template < class T >
+	struct goslice{
+		int Len;
+		int Cap;
+		T *Data; //[Len]
+	};
 }
 
-namespace tmp {
-  struct dict {
-    golang::goslice<double> m_1;
-    golang::goslice<float> m_2;
-    golang::goslice<int> m_3;
-    golang::gostring m_4;
-  };
+namespace {
+	golang::goslice< double> i_1;
+	golang::goslice< float> i_2;
+	golang::goslice< int> i_3;
+	golang::goslice< golang::gostring> i_4;
 }
 
 #endif
